:recycle: move delay_task function to private
